🎨 style(LogInRegistrationForm.vue): refactor media queries for modal width The media queries for the modal width have been refactored to improve readability and consistency. The changes do not affect the functionality of the component. Additionally, the `isSignup` computed property has been renamed to `isNewSignup` to improve clarity. The `console.log` statement has been added to log the first name of the user when signing up. Finally, the confirmation dialog now has a more descriptive content message.
🔥 chore(LogInRegistrationForm.vue): remove empty style tag
The empty style tag was removed from the component as it was not being used and served no purpose.

🐛 fix(UserInfoInput.vue): fix formValue not being reactive
The formValue was not being passed as a reactive object to the onSignUp event, causing the parent component to not receive the updated values. This commit fixes the issue by passing the formValue as a reactive object.