🎨 style(LogInRegistrationForm.vue): refactor showPhoneInput computed property 🔒 chore(LogInRegistrationForm.vue): add useOtpStore hook to import statement The showPhoneInput computed property has been refactored to use the otpStore.otpSent property instead of the showPhoneInput data property. This improves consistency and readability of the code. The useOtpStore hook has been added to the import statement to allow for the use of the otpStore.
🔨 refactor(LogInRegistrationForm.vue): remove unused watch method
🚀 chore(LogInRegistrationForm.vue): add async/await to onSignUp method and update user details
The watch method was removed as it was not being used. The onSignUp method was updated to use async/await and update the user details. The isAuthenticated and hasSignedUp properties are checked to determine if the user has signed up and if so, the user is redirected to the home page.

🐛 fix(LogInRegistrationForm.vue): replace setting user to null with logOut method
✨ feat(LogInRegistrationForm.vue): add styling to heading
The user is now logged out using the logOut method instead of setting the user to null. This ensures that the user is properly logged out and all related data is cleared. The heading now has styling applied to it, which improves the visual appearance of the component.